Pundit

+++ app/controllers/friends_controller.rb
@@ -3,18 +3,20 @@ class FriendsController < ApplicationController
 
   def index
     @friends = current_user.friends
+    @friends = policy_scope(Friend)
+    @users = User.all.where.not(id: current_user.id)
   end
 
   def create
-    friend_full_name = params[:friend_full_name].strip
+    @attendee = User.find(params[:attendee_id])
 
-    friend = User.find_by("LOWER(CONCAT(first_name, ' ', last_name)) = ?", friend_full_name.downcase)
+    authorize(Friend)
 
-    if friend
-      if current_user.friends.exists?(attendee_id: friend.id)
+    if @attendee
+      if current_user.friends.exists?(attendee_id: @attendee.id)
         redirect_to friends_path, notice: "Friend already added."
       else
-        current_user.friends.create(attendee_id: friend.id)
+        current_user.friends.create(attendee_id: @attendee.id)
         redirect_to friends_path, notice: "Friend added successfully."
       end
     else
@@ -22,11 +24,11 @@ class FriendsController < ApplicationController
     end
   end
 
-
-
   def destroy
-    friend = current_user.friends.find_by(attendee_id: params[:id])
+    friend = current_user.friends.find_by(id: params[:id])
+
     if friend
+      authorize(friend)
       friend.destroy
       redirect_to friends_path, notice: "Friend removed successfully."
     else
